fix bwtcompress and infobar screen in morden options, and new read cl…

+++ src/ClassicUO.IO/UOFileIndex.cs
@@ -1,8 +1,8 @@
 ﻿#region license
 
-// Copyright (c) 2021, andreakarasho
+// Copyright (c) 2024, andreakarasho
 // All rights reserved.
-// 
+//
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided that the following conditions are met:
 // 1. Redistributions of source code must retain the above copyright
@@ -16,7 +16,7 @@
 // 4. Neither the name of the copyright holder nor the
 //    names of its contributors may be used to endorse or promote products
 //    derived from this software without specific prior written permission.
-// 
+//
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS ''AS IS'' AND ANY
 // EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
 // WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
@@ -43,8 +43,8 @@ namespace ClassicUO.IO
             long offset,
             int length,
             int decompressed,
-            short width = 0,
-            short height = 0,
+            int width = 0,
+            int height = 0,
             ushort hue = 0
         )
         {
@@ -65,8 +65,8 @@ namespace ClassicUO.IO
         public long Offset;
         public int Length;
         public int DecompressedLength;
-        public short Width;
-        public short Height;
+        public int Width;
+        public int Height;
         public ushort Hue;
         public sbyte AnimOffset;
